+++ R/read_ioc.R
@@ -21,15 +21,15 @@
 #'
 #' path <- "NZIOC_Full_Spreadsheet_December_2021.xlsx"
 #'
-#' nite <- read_nzioc(path)
+#' ioc <- read_ioc(path)
 #' }
 #' @importFrom openxlsx read.xlsx
 #' @export
-read_nzioc <- function(path, version = TRUE) {
+read_ioc <- function(path, version = TRUE) {
 
-  nzioc <- openxlsx::read.xlsx(xlsxFile = path, startRow = 2)
+  ioc <- openxlsx::read.xlsx(xlsxFile = path, startRow = 2)
 
-  names(nzioc) <- c(
+  names(ioc) <- c(
     "cas_number", "cas_name", "approval", "restrictions_exclusions"
   )
 
@@ -45,9 +45,9 @@ read_nzioc <- function(path, version = TRUE) {
     version <- unlist(strsplit(version, split = "\\."))[1]
     version <- rev(unlist(strsplit(version, split = "_")))
     version <- paste(version[2], version[1])
-    nzioc <- transform(nzioc, version = version)
+    ioc <- transform(ioc, version = version)
   }
 
-  nzioc
+  ioc
 
 }